rename gb/log to gb/debug
