+++ build_test.go
@@ -11,12 +11,12 @@ import (
 	"sort"
 	"testing"
 
-	"github.com/constabulary/gb/log"
+	"github.com/constabulary/gb/debug"
 )
 
 func TestBuild(t *testing.T) {
-	log.Verbose = false
-	defer func() { log.Verbose = false }()
+	debug.Verbose = false
+	defer func() { debug.Verbose = false }()
 
 	opts := func(o ...func(*Context) error) []func(*Context) error { return o }
 	tests := []struct {
@@ -97,8 +97,8 @@ func TestBuild(t *testing.T) {
 }
 
 func TestBuildPackage(t *testing.T) {
-	log.Verbose = false
-	defer func() { log.Verbose = false }()
+	debug.Verbose = false
+	defer func() { debug.Verbose = false }()
 	tests := []struct {
 		pkg string
 		err error
@@ -150,8 +150,8 @@ func TestBuildPackage(t *testing.T) {
 }
 
 func TestBuildPackages(t *testing.T) {
-	log.Verbose = false
-	defer func() { log.Verbose = false }()
+	debug.Verbose = false
+	defer func() { debug.Verbose = false }()
 	tests := []struct {
 		pkgs    []string
 		actions []string
