+++ context.go
@@ -15,7 +15,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/constabulary/gb/log"
+	"github.com/constabulary/gb/debug"
 )
 
 // Context represents an execution of one or more Targets inside a Project.
@@ -197,7 +197,7 @@ func (c *Context) ResolvePackage(path string) (*Package, error) {
 
 // Destroy removes the temporary working files of this context.
 func (c *Context) Destroy() error {
-	log.Debugf("removing work directory: %v", c.workdir)
+	debug.Debugf("removing work directory: %v", c.workdir)
 	return os.RemoveAll(c.workdir)
 }
 
@@ -218,7 +218,7 @@ func runOut(output io.Writer, dir string, env []string, command string, args ...
 	cmd.Stdout = output
 	cmd.Stderr = os.Stderr
 	cmd.Env = mergeEnvLists(env, envForDir(cmd.Dir))
-	log.Debugf("cd %s; %s", cmd.Dir, cmd.Args)
+	debug.Debugf("cd %s; %s", cmd.Dir, cmd.Args)
 	err := cmd.Run()
 	return err
 }
@@ -321,7 +321,7 @@ func (c *Context) isCrossCompile() bool {
 }
 
 func matchPackages(c *Context, pattern string) []string {
-	log.Debugf("matchPackages: %v %v", c.srcdirs[0].Root, pattern)
+	debug.Debugf("matchPackages: %v %v", c.srcdirs[0].Root, pattern)
 	match := func(string) bool { return true }
 	treeCanMatch := func(string) bool { return true }
 	if pattern != "all" && pattern != "std" {
