+++ test/test.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"github.com/constabulary/gb"
-	"github.com/constabulary/gb/log"
+	"github.com/constabulary/gb/debug"
 )
 
 // Test returns a Target representing the result of compiling the
@@ -48,7 +48,7 @@ func TestPackages(flags []string, pkgs ...*gb.Package) (*gb.Action, error) {
 	test := gb.Action{
 		Name: fmt.Sprintf("test: %s", strings.Join(names(pkgs), ",")),
 		Run: func() error {
-			log.Debugf("test duration: %v %v", time.Since(t0), pkgs[0].Statistics.String())
+			debug.Debugf("test duration: %v %v", time.Since(t0), pkgs[0].Statistics.String())
 			return nil
 		},
 	}
