+++ test/test_test.go
@@ -10,12 +10,12 @@ import (
 	"time"
 
 	"github.com/constabulary/gb"
-	"github.com/constabulary/gb/log"
+	"github.com/constabulary/gb/debug"
 )
 
 func TestTest(t *testing.T) {
-	log.Verbose = false
-	defer func() { log.Verbose = false }()
+	debug.Verbose = false
+	defer func() { debug.Verbose = false }()
 	tests := []struct {
 		pkg      string
 		testArgs []string
